Added cvar `cl_scaleSensWithFov`
When set to non-zero the engine will update mouse sensitivity according to the
current FOV. When the user plays at any FOV the mouse sensitivity will be
changed to preserve the same "mouse feeling" as if the user was playing with
specified sensitivity at 90 degree FOV.

The formula to calculate the sensitivity scaling is

```
  tand(current_fov / 2.0) / tand(90 / 2.0)
= tand(current_fov / 2.0)
```

This cvar overrides the zoom sensitivity value which gets calculated in cgame.

It is set to 0 by default, as it breaks mouse sensitivity configs for players
who play with cg_fov != 90.