Fixed floating-point errors when turning.
The `cl_scaleSensWithFov` option could sometimes cause mouse jitter
after a few minutes of gameplay.

After some testing it was found that the cause was because YAW and PITCH
could accumulate to extreme values which makes the floating-point
precision when performning mouse movements to go away. After a few
minutes of gameplay the floating-point errors would become noticable.

The problem has been there prior to the introduction of
`cl_scaleSensWithFov` but wasn't noticed when it is disabled or not
available. This is probably because that option uses trigonometry on the
mouse calculations which increases the floating-point errors.

The solution is to limit YAW and PITCH to values between -180 and 180
degrees.