Updated internal font
Removed dot in middle of zero (0).  Now it is hard to distinguish
between 0 and O, but it looks better when used in HUD.  And people don't
really need to be able to distinguish them.

Also removed some pixel leakage which appears on big characters (mainly
at the scoreboard) at 16:9, but it is impossible to fix everything
without making other characters look ugly.

Notably the character 'z'.  It has problems with pixel leakage on the
top.  But fixing that would require making the 'j' character get cut at
the bottom, which looks ugly.  So that leakage was not fixed.  There is
some around other characters too, but it is not nearly as obvious as
with 'z'.  The problems with pixel leakage is only observable for big
text rendered at widescreen resolutions.

The pixel leakage is caused by problems in the image scaling.  It causes
pixels from one grid-cell to sometimes leak to another neighboring
grid-cells.