Updated icons, with modified icons from OpenArena
The icons are RED instead of GREY/BLUE to differentiate from the normal
OpenArena logo.  The OpenArena icons, which the new icons are based off,
are licensed under GPL, and where taken from the OpenArena project.

Could not generate the .icns for MAC OSX.  So for those files, the Quake
3 logo from ioquake3 is still used, until a solution is found.