Input buffer for tty console -> more responsive
Previously characters where read one at a time from STDIN for the TTY
console.  This combined with the fact that only one key is handled each
iteration of the event loop made the console feel very sluggish.

Now a buffer is used to make the console a lot more responsive.