Add notify line fading, console type and height
The console type is controlled by the `cl_consoleType` cvar with the
following settings:

- 0 is the default console (also in official OpenArena engine)
- 1 is the default console but the transparency of the bar below the
  console is changed by the `cl_consoleColorAlpha` cvar (also available
  in official OpenArena engine)
- 2 is console with filled background where the `cl_consoleColor*` cvars
  control the color and transparency (also available in official
  OpenArena engine)
- 3 is like 2 but the console will fade in instead of drop down when it
  is activated
- 4 is like 2 but the console will both fade in AND drop down when it
  is activated

The console height can be controlled by the cvar `cl_consoleHeight`
which is a value between 0 and 1 where 0.5 is half the screen (default
setting).

When pressing a console key and holding ALT or SHIFT the console will
take up the full screen independent of the `cl_consoleHeight` cvar.

Also fix bug in tty console line printing and remove redundant console
tab alert blinking>